buffer: tweak container allocation strategy
Before this change we were flushing, draining, then shrinking the
containers. The result was that we were never keeping any memory
allocated across flushes and there was therefore significant
reallocation overhead.

This change shrinks the containers _before_ draining them. This means
that - across flushing periods - we keep around exactly the amount of
memory needed to hold the metrics generated during the last period.

When experiencing steady load  - flush intervals with the same number of
metrics on average - memory churn is now significantly reduced.

When experiencing larger bursts, more memory will be kept allocated to
handle the bursts, and will be deallocated in the next "normal" flushing
period.